refactor(Home) : Container & Presenter Pattern
