+++ src/routes/AppRouter.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-import Home from 'components/Home/Home';
+import HomeContainer from 'components/Home/containers/HomeContainer';
 import SearchList from 'components/SearchList/SearchList';
 import Company from 'components/Company/CompanyContainer';
 import HeaderContainer from 'components/Header/containers/HeaderContainer';
@@ -10,7 +10,7 @@ const AppRouter = () => {
 		<Router>
 			<HeaderContainer component={HeaderContainer} />
 			<Switch>
-				<Route path="/" exact component={Home} />
+				<Route path="/" exact component={HomeContainer} />
 				<Route path="/search-list" component={SearchList} />
 				<Route path="/detail" component={Company} />
 			</Switch>
